// src/types/types.ts
export enum ApplicationCommandCategory {
  Core = 'Core',
  Forge = 'Forge',
  Dev = 'Dev',
}

export enum SchemaType {
  String = 'string',
  Number = 'number',
  Boolean = 'boolean',
  Array = 'array',
  Object = 'object',
}

export enum ItemType {
  FILE = 0,
  DIRECTORY = 1,
  SYMLINK = 2,
  UNKNOWN = 3,
}

export enum RequestMethod {
  GET = 'GET',
  POST = 'POST',
  PUT = 'PUT',
  DELETE = 'DELETE',
  PATCH = 'PATCH',
}

export enum ResponseType {
  TEXT = 'TEXT',
  JSON = 'JSON',
  BUFFER = 'BUFFER',
}
// ...
export interface Details {
  category: ApplicationCommandCategory;
  summary?: string;
  usage?: string;
  examples?: string[];
  cooldown?: number;
}

export interface CommandPermission {
  author: PermissionStrings[];
  client: PermissionStrings[];
}

export interface Precondition {
  run(context: any): boolean | Promise<boolean>;
  fail(context: any): void;
}
// ...
export interface ApplicationCommand<
  T extends ApplicationCommandOptions = ApplicationCommandOptions,
> extends CreateApplicationCommand {
  details: Details;
  preconditions?: Precondition;
  permissions?: CommandPermission;
  options?: T;
  acknowledge?: boolean;
  ephemeral?: boolean;
  dev?: boolean;
  run: (bot: CustomBot, interaction: Interaction, options: GetApplicationCommandOptions<T>) => any;
  autocomplete?: (bot: CustomBot, interaction: Interaction, options: GetApplicationCommandOptions<T>) => any;
}
// ...
export type ApplicationCommandOption = Camelize<DiscordApplicationCommandOption>;
export type ApplicationCommandOptions = ApplicationCommandOption[];
// ...
export type Interaction = typeof bot.transformers.$inferredTypes.interaction;
// ...
export type BuildOptions<T extends ApplicationCommandOptions | undefined> = {
  [Prop in keyof Omit<T, keyof unknown[]> as GetOptionName<T[Prop]>]: GetOptionValue<T[Prop]>;
};

export type GetApplicationCommandOptions<T extends ApplicationCommandOptions> = T extends ApplicationCommandOptions
  ? { [Prop in keyof BuildOptions<T> as Prop]: BuildOptions<T>[Prop] }
  : never;